requirements.txt
beautifulsoup4
cerberus
click
isort
ipdb
pyyaml